requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0

whitenoise==6.2.0  # serving static files

ipdb  # debugging

# sentry_sdk  # prodcution sentry SDK
-e ../sentry-python/
#-e /Users/antonpirker/tmp/sentry-python